Update TF-A to 2.6
Update the TF-A to version 2.6, remove the no longer needed
TF-A patch and the associated ENABLE_STACK_PROTECTOR build
option, and disable the AXP805 regulator setup in TF-A using
SUNXI_SETUP_REGULATORS as a build option.

Not performing the regulator setup in TF-A allows Linux to
boot on this board correctly, because the Ethernet PHY needs
a coordinated bringup of two regulators.

Signed-off-by: Dragan Simic <dsimic@manjaro.org>